Migrate pillar unit tests to pytest
